data-lake: Merge transforming functions with the data-lake tools view
